requirements.txt
pytest>=8.0.0
requests>=2.32.0
playwright>=1.52.0
Appium-Python-Client>=4.0.0
python-dotenv>=1.0.1
allure-pytest>=2.13.5
